fix(web): record cache file checksum
Prior to this, simply by the files existance we assumed the checksum
matched. We could recalculate it every time, but with large files that's
a non-trivial pain. As a comprimise between the concerns of convenience
and safety, we can record the checksum in a dedicated file, and use that
instead.