refactor: extract dir-of-collection functionality
Code to locate the relevant base directory for collection-relative file
operations is scattered around this codebase, and in some places doesn't
actually conform to the expected/correct behaviour.

To improve this situation, this functionality now lives in the new file
misc/collectiondir.jl, and is called in several places. This has the
side-effect of making the overall code slightly more readable :)